rename: parser_new.rs -> format_parser.rs (drop vestigial _new suffix)
The _new suffix dated from Phase D when the typed FormatParser trait was
introduced alongside the original parser::FormatParser/OldFormatParser. With
OldFormatParser deleted (and the whole sink layer gone), there is no "old"
to contrast with — the module IS the format-parser module. Renames the file
+ sweeps all crate::parser_new:: -> crate::format_parser:: refs. Mechanical;
850 tests pass, 124/124 value-equivalent, 4 gates green.
